test(verify): move parallel parse tests to sibling module
Keep cmd_verify.rs under the 1000-line soft cap while preserving
the --parallel 0 clap regression coverage.
